v0.3.1
BUG FIX
Apparently when there were no new posts, the `children` array would show up but be empty (either because of Serde's parsing, or because Reddit simply returns an empty array).
I've now accounted for this by first filtering out the subs that had empty children arrays, and only operating on the rest.